refactor: 拆分设置与运维模块
