+++ static/app.js
@@ -39,7 +39,6 @@ let currentMode = 'rmb';
 let chartPeriod = 'realtime';
 let chart = null;
 const MAX_POINTS = 60;
-const RECENT_OPS_LIMIT = 5;
 const CHART_PERIODS = {
   realtime: { label: '价格走势', minutes: null, limit: 60, live: true },
   '1h': { label: '1小时走势', minutes: 60, limit: 360 },
@@ -177,62 +176,10 @@ let appSettings = {
   risk_assistant_cooldown_seconds: 15,
   risk_assistant_cache_minutes: 10,
 };
-let pendingSettingsSave = false;
-let settingsSaveFailed = false;
-let settingsSaveTimer = null;
-const SETTINGS_TABS = ['general', 'email', 'webhook', 'digest', 'risk', 'ops'];
-const SETTINGS_TAB_STORAGE_KEY = 'goldmonitor.settings.activeTab';
-const SETTINGS_TAB_LABELS = {
-  general: '通用设置',
-  email: '邮件通知',
-  webhook: 'Webhook',
-  digest: '摘要通知',
-  risk: '风险分析',
-  ops: '运维与数据',
-};
-const SETTINGS_FIELD_IDS = [
-  'setStartup', 'setStartupTray', 'setFloatingPrice', 'setFloatingDisplayMode',
-  'setFloatingPreset', 'setFloatingOpacity', 'setFloatingSnapEdge', 'setFloatingAlwaysOnTop',
-  'setCloseBehavior', 'setAlertSound', 'setAlertDialog', 'setAlertCooldownMinutes',
-  'setAlertQuietStart', 'setAlertQuietEnd', 'setSmtpServer', 'setSmtpPort',
-  'setSmtpEncryption', 'setSmtpSender', 'setSmtpPassword', 'clearSmtpPassword',
-  'setSmtpRecipient', 'setEmailSubjectTemplate', 'setEmailBodyTemplate', 'setWebhookEnabled',
-  'setWebhookUrl', 'setWebhookWarning', 'setWebhookCritical', 'setWebhookVolatility',
-  'setDailyDigestEnabled', 'setDailyDigestTime', 'setDailyDigestEmail', 'setDailyDigestWebhook',
-  'setRiskAssistantEnabled', 'setRiskAssistantProvider', 'setRiskAssistantDepth',
-  'setDeepseekBaseUrl', 'setDeepseekModel', 'setDeepseekApiKey', 'clearDeepseekApiKey',
-  'setOpenaiCompatibleBaseUrl', 'setOpenaiCompatibleModel', 'setOpenaiCompatibleApiKey',
-  'clearOpenaiCompatibleApiKey', 'setRiskMaxTokens', 'setRiskCooldownSeconds',
-  'setRiskCacheMinutes', 'setExportDir',
-];
-let settingsInitialSnapshot = '';
-let settingsDirty = false;
-let settingsLastFocused = null;
-let activeSettingsTab = 'general';
-let pendingUpdateInfo = null;
-let pendingConfigImportPayload = null;
-let pendingConfigImportPreview = null;
-let configImportPreviewRequestPayload = null;
-let pendingDataArchiveRestore = null;
-let recentOpsRecords = [];
-let autoUpdateTimer = null;
-let lastAutoUpdateCheckAt = 0;
-let opsUpdateStatus = null;
-let onboardingStep = 1;
-let onboardingManual = false;
-let onboardingAutoChecked = false;
-const AUTO_UPDATE_CHECK_INTERVAL_MS = 6 * 60 * 60 * 1000;
-function autoUpdateIntervalMs() {
-  return AUTO_UPDATE_CHECK_INTERVAL_MS;
-}
 let alertEntries = [];
 let alertLogSearch = '';
 let activeAlert = null;
 let mergedAlertCount = 0;
-let deepseekModelOptions = ['deepseek-v4-pro', 'deepseek-v4-flash', 'deepseek-chat', 'deepseek-reasoner'];
-let latestSourceHealthState = { items: [], summary: {} };
-let latestSourceComparisonState = { items: [], summary: {}, status: 'insufficient' };
-let dailyDigestStatusState = {};
 
 function chartUnitLabel() {
   return currentMode === 'usd' ? 'USD/oz' : 'RMB/克';
@@ -859,288 +806,15 @@ socket.on('portfolio_import_undo_error', data => {
   }
 });
 
-socket.on('settings_updated', data => {
-  if (settingsSaveTimer) {
-    clearTimeout(settingsSaveTimer);
-    settingsSaveTimer = null;
-  }
-  const shouldClearProfileMatch = alertProfileSettingsChanged(data || {});
-  if (settingsSaveFailed) {
-    appSettings = Object.assign({}, appSettings, data || {});
-    if (shouldClearProfileMatch) clearCurrentAlertProfileMatch();
-    pendingSettingsSave = false;
-    setSettingsSaving(false);
-    return;
-  }
-  applySettings(data || {});
-  if (shouldClearProfileMatch) clearCurrentAlertProfileMatch();
-  pendingSettingsSave = false;
-  resetSettingsDirtySnapshot();
-  showSettingsMessage('设置已保存并生效。', 'ok');
-});
-
-socket.on('settings_error', data => {
-  if (settingsSaveTimer) {
-    clearTimeout(settingsSaveTimer);
-    settingsSaveTimer = null;
-  }
-  pendingSettingsSave = false;
-  settingsSaveFailed = true;
-  setSettingsDirty(true);
-  showSettingsMessage(data.message || '设置保存失败。', 'error');
-  if (data && data.export_dir_check) renderExportDirStatus(data.export_dir_check);
-});
-
-socket.on('onboarding_started', data => {
-  if (data && data.settings) appSettings = Object.assign({}, appSettings, data.settings);
-});
-
-socket.on('onboarding_completed', data => {
-  const finishButton = document.getElementById('onboardingFinishButton');
-  const skipButton = document.getElementById('onboardingSkipButton');
-  if (finishButton) finishButton.disabled = false;
-  if (skipButton) skipButton.disabled = false;
-  if (!data || data.ok === false) return;
-  if (data.settings) applySettings(data.settings);
-  document.getElementById('onboardingBackdrop').classList.remove('show');
-  if (data.startup_error) setOpsStatus('首次使用设置已保存，但开机自启动设置失败，请检查系统权限。', false);
-});
-
-socket.on('onboarding_error', data => {
-  const message = document.getElementById('onboardingMessage');
-  const finishButton = document.getElementById('onboardingFinishButton');
-  const skipButton = document.getElementById('onboardingSkipButton');
-  if (message) message.textContent = (data && data.message) || '首次使用设置保存失败。';
-  if (finishButton) finishButton.disabled = false;
-  if (skipButton) skipButton.disabled = false;
-});
-
-socket.on('daily_digest_status', data => {
-  applyDailyDigestStatus(data || {});
-});
-
-socket.on('daily_digest_previewed', data => {
-  const button = document.getElementById('btnPreviewDailyDigest');
-  if (button) button.disabled = false;
-  if (!data || data.ok === false) {
-    setDailyDigestStatus((data && data.message) || '生成摘要预览失败。', false);
-    return;
-  }
-  renderDailyDigestPreview(data);
-  setDailyDigestStatus('摘要预览已生成，未发送通知。', true);
-});
-
-socket.on('daily_digest_test_result', data => {
-  const button = document.getElementById('btnTestDailyDigest');
-  if (button) button.disabled = false;
-  if (data && data.digest) renderDailyDigestPreview(data.digest);
-  setDailyDigestStatus(
-    data && data.message ? '测试发送：' + data.message : '测试发送已完成。',
-    !!(data && data.ok)
-  );
-});
+registerSettingsSocketHandlers(socket);
 
 socket.on('threshold_error', data => alert(data.message));
 
-socket.on('update_status', data => {
-  applyUpdateStatus(data || {});
-});
+registerOperationsSocketHandlers(socket);
 
 registerRiskAnalysisSocketHandlers(socket);
 
-socket.on('source_health_updated', data => {
-  renderSourceHealth(data || {});
-});
-
-socket.on('market_sources_updated', data => {
-  setSourceManagerStatus(data && data.message ? data.message : '数据源配置已更新。', true);
-});
-
-socket.on('market_sources_error', data => {
-  setSourceManagerStatus(data && data.message ? data.message : '数据源配置更新失败。', false);
-});
-
-socket.on('market_source_retry_result', data => {
-  const pending = data && data.pending;
-  const message = data && data.message ? data.message : (pending ? '正在探测数据源...' : '数据源探测完成。');
-  setSourceManagerStatus(message, pending ? null : !!(data && data.ok));
-  if (data && data.source_health) renderSourceHealth(data.source_health);
-});
-
 registerHistoryReviewSocketHandlers(socket);
-
-socket.on('config_backup_ready', data => {
-  if (!data) return;
-  if (data.ok === false) {
-    addRecentOpsRecord('config_export', data);
-    setOpsExportStatus(data, '配置已导出', '配置导出失败。');
-    return;
-  }
-  if (data.saved_path) {
-    addRecentOpsRecord('config_export', data);
-    setOpsExportStatus(data, '配置已导出', '配置导出失败。');
-    return;
-  }
-  if (!data.content) return;
-  const fallbackData = { ...data, filename: data.filename || 'GoldMonitor-config.json' };
-  downloadText(fallbackData.filename, data.content, 'application/json;charset=utf-8');
-  addRecentOpsRecord('config_export', fallbackData);
-  setOpsExportStatus(fallbackData, '配置已导出', '配置导出失败。');
-});
-
-socket.on('data_archive_exported', data => {
-  addRecentOpsRecord('data_archive_export', data || {});
-  setOpsExportStatus(data || {}, '完整数据归档已创建', '完整数据归档失败。');
-});
-
-socket.on('data_archive_export_error', data => {
-  addRecentOpsRecord('data_archive_export', data || {});
-  setOpsExportStatus(data || {}, '完整数据归档已创建', '完整数据归档失败。');
-});
-
-socket.on('data_archive_restored', data => {
-  if (!data || data.ok === false) return;
-  setOpsStatus(data.message || '完整数据已恢复，正在重新载入界面。', true);
-});
-
-socket.on('diagnostics_ready', data => {
-  if (!data) return;
-  if (data.ok === false) {
-    addRecentOpsRecord('diagnostics_export', data);
-    setOpsExportStatus(data, '诊断报告已导出', '诊断报告导出失败。');
-    return;
-  }
-  if (data.saved_path) {
-    addRecentOpsRecord('diagnostics_export', data);
-    setOpsExportStatus(data, '诊断报告已导出', '诊断报告导出失败。');
-    return;
-  }
-  if (!data.content) return;
-  const fallbackData = { ...data, filename: data.filename || 'GoldMonitor-diagnostics.json' };
-  downloadText(fallbackData.filename, data.content, 'application/json;charset=utf-8');
-  addRecentOpsRecord('diagnostics_export', fallbackData);
-  setOpsExportStatus(fallbackData, '诊断报告已导出', '诊断报告导出失败。');
-});
-
-function hideDiagnosticsCopyFallback() {
-  const fallback = document.getElementById('diagnosticsCopyFallback');
-  if (!fallback) return;
-  fallback.value = '';
-  fallback.hidden = true;
-}
-
-function showDiagnosticsCopyFallback(content) {
-  const fallback = document.getElementById('diagnosticsCopyFallback');
-  if (!fallback) return;
-  fallback.value = content;
-  fallback.hidden = false;
-  requestAnimationFrame(() => {
-    fallback.focus();
-    fallback.select();
-  });
-}
-
-function copyTextWithSelection(content) {
-  const textarea = document.createElement('textarea');
-  textarea.value = content;
-  textarea.setAttribute('readonly', 'readonly');
-  textarea.style.position = 'fixed';
-  textarea.style.left = '-9999px';
-  textarea.style.top = '0';
-  document.body.appendChild(textarea);
-  textarea.focus();
-  textarea.select();
-  try {
-    return document.execCommand('copy');
-  } catch (error) {
-    return false;
-  } finally {
-    textarea.remove();
-  }
-}
-
-function copyTextToClipboard(content) {
-  const fallbackCopy = () => Promise.resolve(copyTextWithSelection(content));
-  if (navigator.clipboard && navigator.clipboard.writeText) {
-    return navigator.clipboard.writeText(content).then(() => true).catch(fallbackCopy);
-  }
-  return fallbackCopy();
-}
-
-socket.on('diagnostics_copy_ready', data => {
-  if (!data) return;
-  if (data.ok === false) {
-    setOpsStatus(data.message || '诊断摘要生成失败。', false);
-    return;
-  }
-  const content = data.content || '';
-  if (!content) {
-    setOpsStatus('诊断摘要为空，无法复制。', false);
-    return;
-  }
-  copyTextToClipboard(content)
-    .then(copied => {
-      if (copied) {
-        hideDiagnosticsCopyFallback();
-        setOpsStatus('诊断摘要已复制。', true);
-        return;
-      }
-      showDiagnosticsCopyFallback(content);
-      setOpsStatus('自动复制失败，已展示诊断摘要，可手动复制。', false);
-    })
-    .catch(() => {
-      showDiagnosticsCopyFallback(content);
-      setOpsStatus('自动复制失败，已展示诊断摘要，可手动复制。', false);
-    });
-});
-
-socket.on('exports_folder_opened', data => {
-  addRecentOpsRecord('open_exports_folder', data || {});
-  if (data && data.ok === false) {
-    setOpsExportStatus(data, '已打开导出目录', '无法打开导出目录。');
-    return;
-  }
-  setOpsStatus(data && data.message ? data.message : '已打开导出目录。', !!(data && data.ok));
-});
-
-socket.on('config_import_previewed', data => {
-  const text = document.getElementById('configImportText') ? document.getElementById('configImportText').value.trim() : '';
-  const previewedPayload = configImportPreviewRequestPayload;
-  configImportPreviewRequestPayload = null;
-  if (!previewedPayload || previewedPayload !== text) {
-    pendingConfigImportPayload = null;
-    pendingConfigImportPreview = null;
-    setOpsStatus('备份内容已变更，请重新预检。', false);
-    return;
-  }
-  if (data && data.importable) {
-    pendingConfigImportPayload = previewedPayload;
-    pendingConfigImportPreview = data;
-    setOpsStatus(renderConfigImportPreview(data), true);
-    return;
-  }
-  pendingConfigImportPayload = null;
-  pendingConfigImportPreview = null;
-  setOpsStatus(renderConfigImportPreview(data), false);
-});
-
-socket.on('config_import_result', data => {
-  configImportPreviewRequestPayload = null;
-  pendingConfigImportPayload = null;
-  pendingConfigImportPreview = null;
-  setOpsStatus(data && data.message ? data.message : '配置导入完成。', !!(data && data.ok));
-});
-
-socket.on('settings_reset_result', data => {
-  setOpsStatus(data && data.message ? data.message : '已恢复默认设置。', !!(data && data.ok));
-});
-
-socket.on('test_alert_result', data => {
-  const statusEl = document.getElementById('testAlertStatus');
-  if (!statusEl) return;
-  statusEl.textContent = data && data.message ? data.message : '测试提醒已触发。';
-  statusEl.className = 'test-email-status ' + (data && data.ok ? 'ok' : 'fail');
-});
 
 function downloadText(filename, content, mimeType) {
   const blob = new Blob([content], { type: mimeType || 'text/plain;charset=utf-8' });
@@ -1152,1595 +826,6 @@ function downloadText(filename, content, mimeType) {
   link.click();
   link.remove();
   setTimeout(() => URL.revokeObjectURL(url), 2000);
-}
-
-function toggleSourceHealthDetails() {
-  const details = document.getElementById('sourceHealthDetails');
-  if (!details) return;
-  details.hidden = !details.hidden;
-}
-
-function sourceQualityText(quality) {
-  if (!quality || typeof quality !== 'object') return '';
-  const score = quality.score == null ? '--' : quality.score;
-  const label = quality.label || quality.level || '--';
-  return '行情质量 ' + score + '分/' + label;
-}
-
-function setSourceManagerStatus(message, ok) {
-  const status = document.getElementById('sourceManagerStatus');
-  if (!status) return;
-  status.textContent = message || '';
-  status.className = 'source-manager-status' + (ok === true ? ' ok' : ok === false ? ' fail' : '');
-}
-
-function renderMarketQualityDetails(quality) {
-  const box = document.getElementById('marketQualityDetails');
-  if (!box) return;
-  const reasons = box.querySelector('.market-quality-reasons');
-  const deductions = quality && Array.isArray(quality.deductions) ? quality.deductions : [];
-  if (!deductions.length) {
-    reasons.innerHTML = '<div class="market-quality-reason none"><span class="market-quality-points">0分</span><span>当前没有质量扣分项</span></div>';
-    return;
-  }
-  reasons.innerHTML = deductions.map(item => [
-    '<div class="market-quality-reason" title="' + escapeHtml(item.detail || item.label || '') + '">',
-    '<span class="market-quality-points">-' + escapeHtml(item.points == null ? '--' : item.points) + '分</span>',
-    '<span>' + escapeHtml(item.detail || item.label || '质量异常') + '</span>',
-    '</div>',
-  ].join('')).join('');
-}
-
-function sourceCategoryLabel(category) {
-  if (category === 'gold') return '金价源';
-  if (category === 'forex') return '汇率源';
-  return category || '数据源';
-}
-
-function marketSourcePreferences() {
-  const adapters = latestSourceHealthState && latestSourceHealthState.adapters || {};
-  const enabled = {};
-  const order = {};
-  Object.keys(adapters).forEach(category => {
-    const items = Array.isArray(adapters[category]) ? adapters[category].slice() : [];
-    items.sort((left, right) => Number(left.order || 0) - Number(right.order || 0));
-    order[category] = items.map(item => item.key).filter(Boolean);
-    enabled[category] = items.filter(item => item.enabled).map(item => item.key).filter(Boolean);
-  });
-  return { enabled, order };
-}
-
-function updateMarketSourceEnabled(category, key, checked) {
-  const preferences = marketSourcePreferences();
-  const categoryEnabled = Array.isArray(preferences.enabled[category]) ? preferences.enabled[category].slice() : [];
-  if (checked && !categoryEnabled.includes(key)) categoryEnabled.push(key);
-  if (!checked) preferences.enabled[category] = categoryEnabled.filter(item => item !== key);
-  else preferences.enabled[category] = preferences.order[category].filter(item => categoryEnabled.includes(item));
-  if (!preferences.enabled[category].length) {
-    setSourceManagerStatus(sourceCategoryLabel(category) + '至少启用一个。', false);
-    renderSourceManager(latestSourceHealthState);
-    return;
-  }
-  setSourceManagerStatus('正在保存数据源配置...', null);
-  socket.emit('update_market_sources', preferences);
-}
-
-function moveMarketSource(category, key, direction) {
-  const preferences = marketSourcePreferences();
-  const order = Array.isArray(preferences.order[category]) ? preferences.order[category].slice() : [];
-  const currentIndex = order.indexOf(key);
-  const nextIndex = currentIndex + Number(direction || 0);
-  if (currentIndex < 0 || nextIndex < 0 || nextIndex >= order.length) return;
-  const displaced = order[nextIndex];
-  order[nextIndex] = key;
-  order[currentIndex] = displaced;
-  preferences.order[category] = order;
-  preferences.enabled[category] = order.filter(item => preferences.enabled[category].includes(item));
-  setSourceManagerStatus('正在保存数据源顺序...', null);
-  socket.emit('update_market_sources', preferences);
-}
-
-function retryMarketSource(key) {
-  setSourceManagerStatus('正在探测数据源...', null);
-  socket.emit('retry_market_source', { key });
-}
-
-function resetMarketSources() {
-  setSourceManagerStatus('正在恢复默认数据源顺序...', null);
-  socket.emit('reset_market_sources');
-}
-
-function renderSourceManager(data) {
-  const box = document.getElementById('sourceManager');
-  if (!box) return;
-  const list = box.querySelector('.source-manager-list');
-  const adapters = data && data.adapters && typeof data.adapters === 'object' ? data.adapters : {};
-  const categories = ['gold', 'forex'].filter(category => Array.isArray(adapters[category]));
-  if (!categories.length) {
-    list.innerHTML = '<div class="source-manager-meta">等待数据源目录</div>';
-    return;
-  }
-  list.innerHTML = categories.map(category => {
-    const items = adapters[category].slice().sort((left, right) => Number(left.order || 0) - Number(right.order || 0));
-    const enabledCount = items.filter(item => item.enabled).length;
-    const rows = items.map((item, index) => {
-      const successRate = item.success_rate_pct == null ? '--' : Number(item.success_rate_pct).toFixed(1) + '%';
-      const latency = item.median_latency_ms == null ? '--' : Number(item.median_latency_ms).toFixed(0) + 'ms';
-      const failures = Number(item.consecutive_failures || 0);
-      const currentLabel = item.active ? '当前主源' : item.current_cached ? '当前缓存来源' : item.current ? '正在切换' : '';
-      const disableToggle = !!item.enabled && enabledCount <= 1;
-      const safeKey = escapeHtml(item.key || '');
-      const safeCategory = escapeHtml(category);
-      return [
-        '<div class="source-manager-row' + (item.enabled ? '' : ' disabled') + '">',
-        '<input class="source-manager-toggle" type="checkbox" aria-label="启用' + escapeHtml(item.name || '') + '" ',
-        item.enabled ? 'checked ' : '',
-        disableToggle ? 'disabled ' : '',
-        'onchange="updateMarketSourceEnabled(\'' + safeCategory + '\',\'' + safeKey + '\',this.checked)">',
-        '<div class="source-manager-copy">',
-        '<div class="source-manager-name">' + escapeHtml(item.name || '--') + (currentLabel ? '<span class="source-manager-current">' + currentLabel + '</span>' : '') + '</div>',
-        '<div class="source-manager-meta">近 ' + escapeHtml(item.sample_count || 0) + ' 次 · 成功率 ' + escapeHtml(successRate) + ' · 中位延迟 ' + escapeHtml(latency) + (failures ? ' · 连续失败 ' + failures + ' 次' : '') + '</div>',
-        '</div>',
-        '<div class="source-manager-actions">',
-        '<button class="btn-clear-sm btn-muted-sm" type="button" onclick="moveMarketSource(\'' + safeCategory + '\',\'' + safeKey + '\',-1)" ' + (index === 0 ? 'disabled' : '') + '>上移</button>',
-        '<button class="btn-clear-sm btn-muted-sm" type="button" onclick="moveMarketSource(\'' + safeCategory + '\',\'' + safeKey + '\',1)" ' + (index === items.length - 1 ? 'disabled' : '') + '>下移</button>',
-        '<button class="btn-clear-sm btn-muted-sm" type="button" onclick="retryMarketSource(\'' + safeKey + '\')">探测</button>',
-        '</div>',
-        '</div>',
-      ].join('');
-    }).join('');
-    return '<div class="source-manager-category"><div class="source-manager-category-title">' + sourceCategoryLabel(category) + '</div>' + rows + '</div>';
-  }).join('');
-}
-
-function renderSourceHealth(data) {
-  latestSourceHealthState = data || { items: [], summary: {} };
-  if (data && data.comparison) renderSourceComparison(data.comparison);
-  renderMarketQualityDetails(data && data.quality ? data.quality : {});
-  renderSourceManager(latestSourceHealthState);
-  const box = document.getElementById('sourceHealth');
-  if (!box) return;
-  const items = Array.isArray(data.items) ? data.items : [];
-  const summary = data.summary || {};
-  const head = box.querySelector('.source-summary-text');
-  const list = box.querySelector('.source-health-list');
-  const ok = Number(summary.ok || 0);
-  const failed = Number(summary.failed || 0);
-  const cached = Number(summary.cached || 0);
-  const countText = failed
-    ? '异常 ' + failed + ' · 正常 ' + ok
-    : (cached ? '缓存 ' + cached + ' · 正常 ' + ok : '正常 ' + ok);
-  head.textContent = [sourceQualityText(data.quality), countText].filter(Boolean).join(' · ');
-  head.title = head.textContent;
-  if (!items.length) {
-    list.innerHTML = '<div class="source-health-item"><span class="source-health-dot"></span><span class="source-health-name">等待数据源检查</span><span class="source-health-meta">--</span></div>';
-    return;
-  }
-  list.innerHTML = items.map(item => {
-    const cls = item.cached ? 'cached' : item.ok ? 'ok' : 'fail';
-    const elapsed = item.elapsed_ms == null ? '--' : item.elapsed_ms + 'ms';
-    const status = item.cached ? '缓存' : item.ok ? '正常' : '异常';
-    const title = item.error ? item.error : status;
-    const successRate = item.success_rate_pct == null ? '--' : Number(item.success_rate_pct).toFixed(1) + '%';
-    const failures = Number(item.consecutive_failures || 0);
-    const rolling = '成功率 ' + successRate + ' · ' + elapsed + (failures ? ' · 连续失败 ' + failures + ' 次' : '');
-    return [
-      '<div class="source-health-item" title="' + escapeHtml(title) + '">',
-      '<span class="source-health-dot ' + cls + '"></span>',
-      '<span class="source-health-name">' + escapeHtml(item.name || '--') + (item.active ? ' · 当前主源' : '') + '</span>',
-      '<span class="source-health-meta">' + escapeHtml(status + ' · ' + rolling) + '</span>',
-      '</div>',
-    ].join('');
-  }).join('');
-}
-
-function renderSourceComparison(data) {
-  latestSourceComparisonState = Object.assign({ items: [], summary: {}, status: 'insufficient' }, data || {});
-  const box = document.getElementById('sourceComparison');
-  if (!box) return;
-  const head = box.querySelector('.source-comparison-head span:first-child');
-  const badge = box.querySelector('.source-comparison-badge');
-  const list = box.querySelector('.source-comparison-list');
-  const summary = latestSourceComparisonState.summary || {};
-  const status = latestSourceComparisonState.status || 'insufficient';
-  const statusText = status === 'anomaly' ? '异常' : status === 'normal' ? '正常' : '不足';
-  head.textContent = summary.spread_pct == null
-    ? '行情源价差'
-    : '行情源价差 ' + Number(summary.spread_pct).toFixed(2) + '%';
-  badge.textContent = statusText;
-  badge.className = 'source-comparison-badge ' + status;
-  const items = Array.isArray(latestSourceComparisonState.items)
-    ? latestSourceComparisonState.items.filter(item => item && item.usd != null).slice(0, 4)
-    : [];
-  if (!items.length) {
-    list.innerHTML = '<div class="source-comparison-item"><span class="source-comparison-name">等待行情源样本</span><span class="source-comparison-price">--</span></div>';
-    return;
-  }
-  list.innerHTML = items.map(item => {
-    const state = item.cached ? '缓存' : item.stale ? '过期' : item.available ? '可比' : '待确认';
-    return [
-      '<div class="source-comparison-item" title="' + escapeHtml((item.name || '') + ' · ' + state) + '">',
-      '<span class="source-comparison-name">' + escapeHtml(item.name || '--') + ' · ' + escapeHtml(state) + '</span>',
-      '<span class="source-comparison-price">$' + Number(item.usd).toFixed(2) + '</span>',
-      '</div>',
-    ].join('');
-  }).join('');
-}
-
-function refreshSourceHealth() {
-  socket.emit('get_source_health');
-}
-
-
-function setOpsStatus(message, ok) {
-  const el = document.getElementById('opsStatus');
-  if (!el) return;
-  el.textContent = message || '';
-  el.style.color = ok ? 'var(--down)' : 'var(--up)';
-}
-
-function recentOpsTypeLabel(type) {
-  if (type === 'config_export') return '导出配置';
-  if (type === 'data_archive_export') return '完整数据归档';
-  if (type === 'diagnostics_export') return '生成诊断';
-  if (type === 'open_exports_folder') return '打开目录';
-  return '运维操作';
-}
-
-function recentOpsTimeLabel(date) {
-  return date.toLocaleTimeString('zh-CN', { hour12: false, hour: '2-digit', minute: '2-digit', second: '2-digit' });
-}
-
-function addRecentOpsRecord(type, data) {
-  const payload = data && typeof data === 'object' ? data : {};
-  const detail = payload.error_detail && typeof payload.error_detail === 'object' ? payload.error_detail : {};
-  const dirCheck = payload.export_dir_check && typeof payload.export_dir_check === 'object' ? payload.export_dir_check : {};
-  const ok = payload.ok !== false;
-  const savedPath = payload.saved_path || '';
-  const path = savedPath || payload.export_dir || detail.export_dir || dirCheck.path || '';
-  const message = ok
-    ? (savedPath ? '文件已保存到导出目录。' : (type === 'open_exports_folder' ? '已打开导出目录。' : (payload.message || '操作完成。')))
-    : (detail.message || payload.message || '操作失败。');
-  const record = {
-    id: Date.now() + '-' + recentOpsRecords.length,
-    type,
-    label: recentOpsTypeLabel(type),
-    ok,
-    status: ok ? '成功' : '失败',
-    time: recentOpsTimeLabel(new Date()),
-    path,
-    message,
-    error: detail.error || '',
-    actions: !ok && Array.isArray(dirCheck.actions) ? dirCheck.actions : [],
-  };
-  recentOpsRecords = [record, ...recentOpsRecords].slice(0, RECENT_OPS_LIMIT);
-  renderRecentOpsRecords();
-}
-
-function renderRecentOpsRecords() {
-  const list = document.getElementById('recentOpsList');
-  if (!list) return;
-  if (!recentOpsRecords.length) {
-    list.innerHTML = '<div class="ops-recent-empty">暂无操作记录</div>';
-    return;
-  }
-  list.innerHTML = recentOpsRecords.map(record => {
-    const stateClass = record.ok ? 'ok' : 'fail';
-    const actions = Array.isArray(record.actions) ? record.actions.map(exportDirActionButton).filter(Boolean).join('') : '';
-    const path = record.path ? '<div class="ops-recent-path" title="' + escapeHtml(record.path) + '">' + escapeHtml(record.path) + '</div>' : '';
-    const failure = !record.ok
-      ? '<div class="ops-recent-error"><strong>失败原因</strong><span>' + escapeHtml([record.message, record.error ? '底层错误：' + record.error : ''].filter(Boolean).join(' ')) + '</span></div>'
-      : '';
-    return [
-      '<div class="ops-recent-item ' + stateClass + '">',
-      '<div class="ops-recent-head">',
-      '<span class="ops-recent-title">' + escapeHtml(record.label) + '</span>',
-      '<span class="ops-recent-state ' + stateClass + '">' + escapeHtml(record.status) + '</span>',
-      '<span class="ops-recent-time">' + escapeHtml(record.time) + '</span>',
-      '</div>',
-      path,
-      record.ok ? '<div class="ops-recent-message" title="' + escapeHtml(record.message) + '">' + escapeHtml(record.message) + '</div>' : failure,
-      actions ? '<div class="export-dir-actions">' + actions + '</div>' : '',
-      '</div>',
-    ].join('');
-  }).join('');
-}
-
-function setOpsExportStatus(data, successLabel, fallbackMessage) {
-  const el = document.getElementById('opsStatus');
-  if (!el) return;
-  const payload = data && typeof data === 'object' ? data : {};
-  const ok = payload.ok !== false;
-  el.style.color = ok ? 'var(--down)' : 'var(--up)';
-  if (ok) {
-    const savedPath = payload.saved_path || '';
-    const filename = payload.filename || '';
-    let message = successLabel || '导出已完成';
-    if (savedPath) {
-      message += '：' + savedPath;
-    } else if (filename) {
-      message += '，文件名：' + filename + '。';
-    } else {
-      message += '。';
-    }
-    el.innerHTML = [
-      '<span>' + escapeHtml(message) + '</span>',
-      savedPath ? '<button class="btn-clear-sm btn-muted-sm export-dir-action" type="button" onclick="openExportsFolder()">打开目录</button>' : '',
-    ].join('');
-    return;
-  }
-  const detail = data && data.error_detail && typeof data.error_detail === 'object' ? data.error_detail : {};
-  const dirCheck = data && data.export_dir_check && typeof data.export_dir_check === 'object' ? data.export_dir_check : {};
-  const message = detail.message || payload.message || fallbackMessage || '导出失败。';
-  const extra = dirCheck.message && dirCheck.message !== message ? dirCheck.message : '';
-  const error = detail.error ? '底层错误：' + detail.error : '';
-  const actions = Array.isArray(dirCheck.actions) ? dirCheck.actions.map(exportDirActionButton).filter(Boolean) : [];
-  el.innerHTML = [
-    '<span>' + escapeHtml([message, extra, error].filter(Boolean).join(' ')) + '</span>',
-    actions.length ? '<span class="export-dir-actions">' + actions.join('') + '</span>' : '',
-  ].join('');
-}
-
-function exportConfig() {
-  setOpsStatus('正在导出配置...', true);
-  socket.emit('export_config');
-}
-
-function exportDataArchive() {
-  setOpsStatus('正在创建完整数据归档...', true);
-  socket.emit('export_data_archive');
-}
-
-function formatDataArchiveBytes(value) {
-  const bytes = Math.max(0, Number(value) || 0);
-  if (bytes < 1024) return bytes + ' B';
-  if (bytes < 1024 * 1024) return (bytes / 1024).toFixed(1) + ' KB';
-  return (bytes / (1024 * 1024)).toFixed(1) + ' MB';
-}
-
-function resetDataArchivePreview() {
-  pendingDataArchiveRestore = null;
-  const preview = document.getElementById('dataArchivePreview');
-  const restoreButton = document.getElementById('restoreDataArchiveButton');
-  if (preview) preview.innerHTML = '';
-  if (restoreButton) {
-    restoreButton.hidden = true;
-    restoreButton.disabled = false;
-  }
-}
-
-function renderDataArchivePreview(data, fileName) {
-  const preview = document.getElementById('dataArchivePreview');
-  const restoreButton = document.getElementById('restoreDataArchiveButton');
-  if (!preview || !restoreButton) return;
-  if (!data || data.ok === false || data.restorable === false) {
-    preview.innerHTML = '<div class="data-archive-preview-error">' + escapeHtml((data && data.message) || '归档预检失败。') + '</div>';
-    restoreButton.hidden = true;
-    return;
-  }
-  const items = Array.isArray(data.items) ? data.items.filter(item => item && item.present) : [];
-  const labels = items.slice(0, 8).map(item => item.label || item.key).filter(Boolean);
-  const remaining = Math.max(0, items.length - labels.length);
-  const detail = labels.join('、') + (remaining ? '等 ' + items.length + ' 项' : '');
-  preview.innerHTML = [
-    '<div class="data-archive-preview-ok"><strong>归档校验通过</strong></div>',
-    '<div>文件：' + escapeHtml(fileName || '') + '</div>',
-    '<div>来源版本：' + escapeHtml(data.source_app_version || '未知') + '；导出时间：' + escapeHtml((data.exported_at || '').replace('T', ' ')) + '</div>',
-    '<div>数据量：' + escapeHtml(String(data.files || 0)) + ' 项，' + escapeHtml(formatDataArchiveBytes(data.bytes)) + '</div>',
-    detail ? '<div>包含：' + escapeHtml(detail) + '</div>' : '',
-    data.contains_sensitive_data ? '<div class="data-archive-preview-warning">归档包含通知密钥等敏感配置。</div>' : '',
-  ].join('');
-  restoreButton.hidden = false;
-}
-
-function chooseDataArchive() {
-  const input = document.getElementById('dataArchiveFile');
-  if (!input) {
-    setOpsStatus('未找到归档文件选择入口。', false);
-    return;
-  }
-  input.value = '';
-  resetDataArchivePreview();
-  input.click();
-}
-
-async function previewDataArchive(input) {
-  const file = input && input.files ? input.files[0] : null;
-  resetDataArchivePreview();
-  if (!file) return;
-  setOpsStatus('正在校验完整数据归档...', true);
-  const formData = new FormData();
-  formData.append('archive', file, file.name);
-  try {
-    const response = await fetch('/api/data-archive/preview', {
-      method: 'POST',
-      headers: { 'X-GoldMonitor-Token': SOCKET_ACCESS_TOKEN },
-      body: formData,
-    });
-    const data = await response.json();
-    renderDataArchivePreview(data, file.name);
-    if (!response.ok || !data.restore_token) {
-      setOpsStatus(data.message || '归档预检失败。', false);
-      return;
-    }
-    pendingDataArchiveRestore = {
-      token: data.restore_token,
-      fileName: file.name,
-      preview: data,
-    };
-    setOpsStatus(data.message || '归档预检通过，请确认恢复。', true);
-  } catch (error) {
-    renderDataArchivePreview({ ok: false, message: '无法上传或校验归档文件。' }, file.name);
-    setOpsStatus('无法上传或校验归档文件。', false);
-  }
-}
-
-async function confirmDataArchiveRestore() {
-  if (!pendingDataArchiveRestore || !pendingDataArchiveRestore.token) {
-    setOpsStatus('请先选择并校验归档文件。', false);
-    return;
-  }
-  const confirmed = confirm('恢复完整数据会覆盖当前设置、持仓、预警、复盘和历史记录。是否继续？');
-  if (!confirmed) return;
-  const restoreButton = document.getElementById('restoreDataArchiveButton');
-  if (restoreButton) restoreButton.disabled = true;
-  setOpsStatus('正在恢复完整数据，请勿关闭应用...', true);
-  try {
-    const response = await fetch('/api/data-archive/restore', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        'X-GoldMonitor-Token': SOCKET_ACCESS_TOKEN,
-      },
-      body: JSON.stringify({ restore_token: pendingDataArchiveRestore.token }),
-    });
-    const data = await response.json();
-    pendingDataArchiveRestore = null;
-    if (!response.ok || data.ok === false) {
-      setOpsStatus(data.message || '完整数据恢复失败，原数据已回滚。', false);
-      if (restoreButton) restoreButton.disabled = false;
-      return;
-    }
-    setOpsStatus(data.message || '完整数据已恢复，正在重新载入界面。', true);
-    setTimeout(() => window.location.reload(), 800);
-  } catch (error) {
-    setOpsStatus('完整数据恢复请求失败，请重新选择归档并确认当前数据状态。', false);
-    if (restoreButton) restoreButton.disabled = false;
-  }
-}
-
-function configImportSectionLabel(section) {
-  if (section === 'settings') return '通用设置';
-  if (section === 'thresholds') return '预警阈值';
-  if (section === 'alert_profiles') return '预警策略模板';
-  if (section === 'alert_rules') return '统一预警规则';
-  return section || '未知配置';
-}
-
-function configImportSecretActionLabel(action) {
-  if (action === 'import') return '导入';
-  if (action === 'clear') return '清空';
-  return '保留现有';
-}
-
-function configImportFormatText(data) {
-  const schemaVersion = Number(data && data.schema_version);
-  const expectedSchemaVersion = Number(data && data.expected_schema_version);
-  const format = data && typeof data.format === 'string' ? data.format.trim() : '';
-  const sourceAppVersion = data && typeof data.source_app_version === 'string'
-    ? data.source_app_version.trim()
-    : '';
-  if (data && data.needs_migration) {
-    return '旧版备份将在导入时迁移' + (sourceAppVersion ? '（来源版本 ' + sourceAppVersion + '）' : '');
-  }
-  const resolvedVersion = Number.isInteger(schemaVersion) && schemaVersion >= 0
-    ? schemaVersion
-    : (Number.isInteger(expectedSchemaVersion) && expectedSchemaVersion >= 0 ? expectedSchemaVersion : null);
-  const formatText = resolvedVersion !== null ? 'schema v' + resolvedVersion : (format || '当前格式');
-  return '当前备份格式：' + formatText + (sourceAppVersion ? '（来源版本 ' + sourceAppVersion + '）' : '');
-}
-
-function renderConfigImportPreview(data) {
-  if (!data || data.ok === false || data.importable === false) {
-    return (data && data.message) || '配置导入预检失败。';
-  }
-  const rawSections = Array.isArray(data.sections) ? data.sections : [];
-  const sections = rawSections.map(configImportSectionLabel);
-  const ignored = data.ignored && typeof data.ignored === 'object' ? data.ignored : {};
-  const ignoredFieldCount = []
-    .concat(Array.isArray(ignored.settings) ? ignored.settings : [])
-    .concat(Array.isArray(ignored.thresholds) ? ignored.thresholds : [])
-    .length;
-  const ignoredProfileCount = Array.isArray(ignored.alert_profiles) ? ignored.alert_profiles.length : 0;
-  const ignoredRuleCount = Array.isArray(ignored.alert_rules) ? ignored.alert_rules.length : 0;
-  const secretActions = rawSections.includes('settings') && data.secret_actions && typeof data.secret_actions === 'object'
-    ? data.secret_actions
-    : {};
-  const secretSummary = Object.keys(secretActions).reduce((acc, key) => {
-    const label = configImportSecretActionLabel(secretActions[key]);
-    acc[label] = (acc[label] || 0) + 1;
-    return acc;
-  }, {});
-  const secretText = Object.keys(secretSummary).map(label => label + ' ' + secretSummary[label] + ' 项').join('，');
-  const parts = [
-    '配置预检通过：将导入' + (sections.length ? sections.join('、') : '配置'),
-    configImportFormatText(data),
-  ];
-  if (ignoredFieldCount) parts.push('忽略不支持字段 ' + ignoredFieldCount + ' 项');
-  if (ignoredProfileCount) parts.push('忽略重复、无效或超限策略模板 ' + ignoredProfileCount + ' 项');
-  if (ignoredRuleCount) parts.push('忽略重复、无效或超限预警规则 ' + ignoredRuleCount + ' 项');
-  if (secretText) parts.push('敏感字段：' + secretText);
-  parts.push('再次点击导入确认');
-  return parts.join('；') + '。';
-}
-
-function importConfig() {
-  const text = document.getElementById('configImportText').value.trim();
-  if (!text) {
-    setOpsStatus('请先粘贴配置备份 JSON。', false);
-    return;
-  }
-  if (configImportPreviewRequestPayload !== null) {
-    const changed = configImportPreviewRequestPayload !== text;
-    setOpsStatus(changed ? '备份内容已变更，当前预检返回后请重新预检。' : '正在预检导入配置...', !changed);
-    return;
-  }
-  if (pendingConfigImportPayload === text && pendingConfigImportPreview && pendingConfigImportPreview.importable) {
-    setOpsStatus('正在导入配置...', true);
-    socket.emit('import_config', { payload: text });
-    pendingConfigImportPayload = null;
-    pendingConfigImportPreview = null;
-    return;
-  }
-  pendingConfigImportPayload = null;
-  pendingConfigImportPreview = null;
-  configImportPreviewRequestPayload = text;
-  setOpsStatus('正在预检导入配置...', true);
-  socket.emit('preview_import_config', { payload: text });
-}
-
-function invalidateConfigImportPreviewOnInput() {
-  const hasPreviewState = configImportPreviewRequestPayload !== null
-    || pendingConfigImportPayload !== null
-    || pendingConfigImportPreview !== null;
-  if (!hasPreviewState) return;
-  configImportPreviewRequestPayload = null;
-  pendingConfigImportPayload = null;
-  pendingConfigImportPreview = null;
-  setOpsStatus('备份内容已变更，请重新预检。', false);
-}
-
-const configImportTextInput = document.getElementById('configImportText');
-if (configImportTextInput) {
-  configImportTextInput.addEventListener('input', invalidateConfigImportPreviewOnInput);
-}
-
-function exportDiagnostics() {
-  setOpsStatus('正在生成诊断报告...', true);
-  socket.emit('get_diagnostics');
-}
-
-function copyDiagnostics() {
-  hideDiagnosticsCopyFallback();
-  setOpsStatus('正在生成诊断摘要...', true);
-  socket.emit('copy_diagnostics');
-}
-
-function openExportsFolder() {
-  setOpsStatus('正在打开导出目录...', true);
-  socket.emit('open_exports_folder');
-}
-
-function resetSettings() {
-  if (!confirm('确定恢复默认设置并清空阈值吗？')) return;
-  setOpsStatus('正在恢复默认设置...', true);
-  socket.emit('reset_settings');
-}
-
-function configureSecretClear(inputId, statusId, buttonId, configured, statusText, readyLabel, emptyLabel) {
-  const input = document.getElementById(inputId);
-  const status = document.getElementById(statusId);
-  const button = document.getElementById(buttonId);
-  if (!input || !status || !button) return;
-  input.checked = false;
-  status.textContent = statusText;
-  status.dataset.state = configured ? 'ok' : '';
-  button.disabled = !configured;
-  button.textContent = configured ? readyLabel : emptyLabel;
-  button.classList.remove('marked');
-  button.dataset.defaultLabel = readyLabel;
-  button.dataset.defaultStatus = statusText;
-}
-
-function toggleSecretClear(inputId, statusId, buttonId, selectedStatus) {
-  const input = document.getElementById(inputId);
-  const status = document.getElementById(statusId);
-  const button = document.getElementById(buttonId);
-  if (!input || !status || !button || button.disabled) return;
-  input.checked = !input.checked;
-  if (input.checked) {
-    status.textContent = selectedStatus;
-    status.dataset.state = 'error';
-    button.textContent = '取消删除';
-    button.classList.add('marked');
-    updateSettingsDirtyState();
-    return;
-  }
-  status.textContent = button.dataset.defaultStatus || '';
-  status.dataset.state = button.disabled ? '' : 'ok';
-  button.textContent = button.dataset.defaultLabel || '删除已保存密钥';
-  button.classList.remove('marked');
-  updateSettingsDirtyState();
-}
-
-// ========== 设置 ==========
-function setText(id, text) {
-  const el = document.getElementById(id);
-  if (el) el.textContent = text;
-}
-
-function setRowHidden(id, hidden) {
-  const el = document.getElementById(id);
-  if (el) el.classList.toggle('platform-hidden', !!hidden);
-}
-
-function applyPlatformLabels() {
-  const platform = appSettings.platform || 'windows';
-  const capabilities = appSettings.platform_capabilities || {};
-  const isMac = platform === 'macos';
-  const menuBarMode = capabilities.floating_price_mode === 'menu_bar';
-
-  setText('startupTrayLabel', isMac ? '自启动时进入菜单栏' : '自启动时进入托盘');
-  setText('startupTrayDesc', isMac ? '开机启动后不弹出主窗口，可从菜单栏打开。' : '开机启动后不弹出主窗口，可从右下角托盘打开。');
-  setText('floatingPriceLabel', menuBarMode ? '菜单栏金价' : '桌面金价悬浮条');
-  setText('floatingPriceDesc', menuBarMode ? '在 macOS 菜单栏显示当前金价，并提供显示窗口、刷新和风险分析入口。' : '主窗口隐藏时，仍在桌面右下角显示当前金价。');
-  setText('floatingDisplayDesc', menuBarMode ? '控制菜单栏优先显示人民币、美元或组合价格。' : '控制桌面悬浮条显示人民币、美元或组合内容。');
-  setText('closeChoiceCopy', isMac ? '隐藏到菜单栏后，程序会继续监控金价并在触发条件时提醒。也可以直接退出程序。' : '最小化到右下角托盘后，程序会继续监控金价并在触发条件时提醒。也可以直接退出程序。');
-  setText('closeMinimizeOption', isMac ? '隐藏到菜单栏' : '最小化到托盘');
-  setText('closeMinimizeButton', isMac ? '隐藏到菜单栏' : '最小化到托盘');
-
-  setRowHidden('floatingPresetRow', menuBarMode);
-  setRowHidden('floatingOpacityRow', menuBarMode);
-  setRowHidden('floatingSnapRow', menuBarMode);
-  setRowHidden('floatingTopmostRow', menuBarMode);
-}
-
-function dailyDigestSelectedChannels() {
-  const channels = [];
-  if (appSettings.daily_digest_email_enabled !== false) channels.push('email');
-  if (appSettings.daily_digest_webhook_enabled) channels.push('webhook');
-  return channels;
-}
-
-function dailyDigestChannelText(channels) {
-  const labels = { email: '邮件', webhook: 'Webhook' };
-  const selected = Array.isArray(channels) ? channels : [];
-  return selected.length ? selected.map(channel => labels[channel] || channel).join('、') : '未选择';
-}
-
-function dailyDigestTimestamp(value) {
-  const text = String(value || '').trim();
-  return text ? text.replace('T', ' ').slice(0, 16) : '';
-}
-
-function setDailyDigestStatus(message, ok) {
-  const status = document.getElementById('dailyDigestStatus');
-  if (!status) return;
-  status.textContent = message || '';
-  status.className = 'setting-status setting-status-wide';
-  if (ok === true) status.dataset.state = 'ok';
-  else if (ok === false) status.dataset.state = 'error';
-  else delete status.dataset.state;
-}
-
-function applyDailyDigestStatus(data) {
-  const next = data && typeof data === 'object' ? data : {};
-  dailyDigestStatusState = Object.assign({}, dailyDigestStatusState, next, {
-    state: Object.assign({}, dailyDigestStatusState.state || {}, next.state || {}),
-    schedule: Object.assign({}, dailyDigestStatusState.schedule || {}, next.schedule || {}),
-  });
-  const enabled = dailyDigestStatusState.enabled != null
-    ? !!dailyDigestStatusState.enabled
-    : !!appSettings.daily_digest_enabled;
-  const time = dailyDigestStatusState.time || appSettings.daily_digest_time || '20:00';
-  const channels = Array.isArray(dailyDigestStatusState.channels)
-    ? dailyDigestStatusState.channels
-    : dailyDigestSelectedChannels();
-  const state = dailyDigestStatusState.state || {};
-  const schedule = dailyDigestStatusState.schedule || {};
-  const parts = [enabled ? '已启用，每日 ' + time : '未启用', '渠道：' + dailyDigestChannelText(channels)];
-  const completedAt = dailyDigestTimestamp(state.last_completed_at);
-  const testedAt = dailyDigestTimestamp(state.last_test_at);
-  if (completedAt) parts.push('最近计划执行：' + completedAt);
-  else if (enabled && schedule.reason === 'before_schedule') parts.push('等待今日计划时间');
-  else if (enabled && schedule.reason === 'due') parts.push('等待任务执行');
-  else if (enabled && schedule.reason === 'invalid_schedule') parts.push('发送时间无效');
-  if (testedAt) parts.push('最近测试：' + testedAt);
-  if (state.last_message) parts.push(state.last_message);
-  const okStatuses = ['sent', 'queued'];
-  const failStatuses = ['failed', 'partial', 'skipped'];
-  const statusFlag = okStatuses.includes(state.last_status)
-    ? true
-    : failStatuses.includes(state.last_status) || schedule.reason === 'invalid_schedule'
-      ? false
-      : null;
-  setDailyDigestStatus(parts.join('；'), statusFlag);
-}
-
-function renderDailyDigestPreview(data) {
-  const preview = document.getElementById('dailyDigestPreview');
-  if (!preview) return;
-  const digest = data && data.digest ? data.digest : (data || {});
-  const subject = String(digest.subject || '').trim();
-  const message = String(digest.message || '').trim();
-  preview.value = [subject, message].filter(Boolean).join('\n\n');
-}
-
-function applySettings(data) {
-  appSettings = Object.assign({}, appSettings, data);
-  applyPlatformLabels();
-  const closeBehavior = appSettings.close_behavior || 'ask';
-  document.getElementById('setStartup').checked = !!appSettings.startup_enabled;
-  document.getElementById('setStartupTray').checked = !!appSettings.startup_to_tray;
-  document.getElementById('setFloatingPrice').checked = appSettings.floating_price_enabled !== false;
-  document.getElementById('setFloatingDisplayMode').value = appSettings.floating_price_display_mode || 'rmb_usd';
-  document.getElementById('setFloatingPreset').value = appSettings.floating_price_preset || 'compact';
-  document.getElementById('setFloatingOpacity').value = appSettings.floating_price_opacity || 94;
-  document.getElementById('setFloatingSnapEdge').checked = appSettings.floating_price_snap_edge !== false;
-  document.getElementById('setFloatingAlwaysOnTop').checked = !!appSettings.floating_price_always_on_top;
-  document.getElementById('setCloseBehavior').value = closeBehavior;
-  document.getElementById('setAlertSound').checked = !!appSettings.alert_sound_enabled;
-  document.getElementById('setAlertDialog').checked = !!appSettings.alert_dialog_enabled;
-  // 邮件通知
-  document.getElementById('setSmtpServer').value = appSettings.smtp_server || '';
-  document.getElementById('setSmtpPort').value = appSettings.smtp_port || '465';
-  document.getElementById('setSmtpEncryption').value = appSettings.smtp_encryption || 'ssl';
-  document.getElementById('setSmtpSender').value = appSettings.smtp_sender || '';
-  document.getElementById('setSmtpPassword').value = '';
-  const smtpPasswordStatus = appSettings.smtp_password_configured
-    ? '已保存授权码：' + (appSettings.smtp_password_masked || '******') + '。输入新授权码后保存会替换当前授权码。'
-    : '未保存授权码。输入授权码后保存即可启用邮件发送。';
-  configureSecretClear(
-    'clearSmtpPassword',
-    'smtpPasswordStatus',
-    'clearSmtpPasswordButton',
-    !!appSettings.smtp_password_configured,
-    smtpPasswordStatus,
-    '删除已保存授权码',
-    '暂无已保存授权码'
-  );
-  document.getElementById('setSmtpRecipient').value = appSettings.smtp_recipient || '';
-  document.getElementById('setAlertCooldownMinutes').value = appSettings.alert_cooldown_minutes ?? 30;
-  document.getElementById('setAlertQuietStart').value = appSettings.alert_quiet_start || '';
-  document.getElementById('setAlertQuietEnd').value = appSettings.alert_quiet_end || '';
-  document.getElementById('setEmailSubjectTemplate').value = appSettings.email_subject_template || '[金价预警·{level}] {title}';
-  document.getElementById('setEmailBodyTemplate').value = appSettings.email_body_template || '';
-  document.getElementById('setWebhookEnabled').checked = !!appSettings.webhook_enabled;
-  document.getElementById('setWebhookUrl').value = appSettings.webhook_url || '';
-  document.getElementById('setWebhookWarning').checked = appSettings.webhook_warning_enabled !== false;
-  document.getElementById('setWebhookCritical').checked = appSettings.webhook_critical_enabled !== false;
-  document.getElementById('setWebhookVolatility').checked = appSettings.webhook_volatility_enabled !== false;
-  document.getElementById('setDailyDigestEnabled').checked = !!appSettings.daily_digest_enabled;
-  document.getElementById('setDailyDigestTime').value = appSettings.daily_digest_time || '20:00';
-  document.getElementById('setDailyDigestEmail').checked = appSettings.daily_digest_email_enabled !== false;
-  document.getElementById('setDailyDigestWebhook').checked = !!appSettings.daily_digest_webhook_enabled;
-  applyDailyDigestStatus(Object.assign({}, dailyDigestStatusState, {
-    enabled: !!appSettings.daily_digest_enabled,
-    time: appSettings.daily_digest_time || '20:00',
-    channels: dailyDigestSelectedChannels(),
-  }));
-  document.getElementById('testEmailStatus').textContent = '';
-  document.getElementById('testEmailStatus').className = 'test-email-status';
-  const webhookStatus = document.getElementById('testWebhookStatus');
-  if (webhookStatus) {
-    webhookStatus.textContent = '';
-    webhookStatus.className = 'test-email-status';
-  }
-  const testAlertStatus = document.getElementById('testAlertStatus');
-  if (testAlertStatus) {
-    testAlertStatus.textContent = '';
-    testAlertStatus.className = 'test-email-status';
-  }
-  document.getElementById('setRiskAssistantEnabled').checked = appSettings.risk_assistant_enabled !== false;
-  document.getElementById('setRiskAssistantProvider').value = appSettings.risk_assistant_provider || 'deepseek';
-  document.getElementById('setRiskAssistantDepth').value = appSettings.risk_assistant_depth || 'standard';
-  document.getElementById('setDeepseekBaseUrl').value = appSettings.deepseek_base_url || 'https://api.deepseek.com';
-  renderDeepseekModelOptions(appSettings.deepseek_model || 'deepseek-v4-pro');
-  document.getElementById('setDeepseekApiKey').value = '';
-  const keyStatus = appSettings.deepseek_api_key_configured
-    ? '已保存密钥：' + (appSettings.deepseek_api_key_masked || '******') + '。输入新 Key 后保存会替换当前密钥。'
-    : '未保存密钥。输入 API Key 后保存即可启用该模型。';
-  configureSecretClear(
-    'clearDeepseekApiKey',
-    'deepseekKeyStatus',
-    'clearDeepseekApiKeyButton',
-    !!appSettings.deepseek_api_key_configured,
-    keyStatus,
-    '删除已保存密钥',
-    '暂无已保存密钥'
-  );
-  document.getElementById('setOpenaiCompatibleBaseUrl').value = appSettings.openai_compatible_base_url || '';
-  document.getElementById('setOpenaiCompatibleModel').value = appSettings.openai_compatible_model || '';
-  document.getElementById('setOpenaiCompatibleApiKey').value = '';
-  const compatibleKeyStatus = appSettings.openai_compatible_api_key_configured
-    ? '已保存密钥：' + (appSettings.openai_compatible_api_key_masked || '******') + '。输入新 Key 后保存会替换当前密钥。'
-    : '未保存密钥。输入 API Key 后保存即可启用该接口。';
-  configureSecretClear(
-    'clearOpenaiCompatibleApiKey',
-    'openaiCompatibleKeyStatus',
-    'clearOpenaiCompatibleApiKeyButton',
-    !!appSettings.openai_compatible_api_key_configured,
-    compatibleKeyStatus,
-    '删除已保存密钥',
-    '暂无已保存密钥'
-  );
-  document.getElementById('setRiskMaxTokens').value = appSettings.risk_assistant_max_tokens || 1200;
-  document.getElementById('setRiskCooldownSeconds').value = appSettings.risk_assistant_cooldown_seconds || 15;
-  document.getElementById('setRiskCacheMinutes').value = appSettings.risk_assistant_cache_minutes ?? 10;
-  applyExportDirSetting();
-  const modelTestStatus = document.getElementById('riskModelTestStatus');
-  if (modelTestStatus) {
-    modelTestStatus.textContent = '';
-    modelTestStatus.className = 'model-test-status';
-  }
-  updateRiskProviderFields();
-  updateRiskButtonState();
-  renderAlertRules();
-  scheduleAutoUpdateCheck();
-}
-
-function applyExportDirSetting() {
-  const input = document.getElementById('setExportDir');
-  if (!input) return;
-  const configured = appSettings.export_dir || '';
-  const effective = appSettings.export_dir_effective || appSettings.export_dir_default || '';
-  input.value = configured;
-  renderExportDirStatus(appSettings.export_dir_check, configured
-    ? '当前导出目录：' + effective
-    : '当前使用默认导出目录：' + (effective || '未记录'));
-}
-
-function exportDirActionButton(action) {
-  if (action === 'choose_export_dir') {
-    return '<button class="btn-clear-sm export-dir-action" type="button" onclick="chooseExportDir()">重新选择</button>';
-  }
-  if (action === 'use_default_export_dir') {
-    return '<button class="btn-clear-sm export-dir-action" type="button" onclick="useDefaultExportDirFromError()">使用默认</button>';
-  }
-  if (action === 'open_export_dir') {
-    return '<button class="btn-clear-sm export-dir-action" type="button" onclick="openExportsFolder()">打开当前目录</button>';
-  }
-  return '';
-}
-
-function renderExportDirStatus(check, fallbackText) {
-  const status = document.getElementById('exportDirStatus');
-  if (!status) return;
-  const data = check && typeof check === 'object' ? check : null;
-  if (!data || !data.status) {
-    status.textContent = fallbackText || '留空使用默认导出目录。';
-    delete status.dataset.state;
-    return;
-  }
-  const statusClass = data.ok ? 'ok' : 'fail';
-  status.dataset.state = data.ok ? 'ok' : 'error';
-  const actions = Array.isArray(data.actions) ? data.actions.map(exportDirActionButton).filter(Boolean) : [];
-  status.innerHTML = [
-    '<span class="export-dir-check ' + statusClass + '">' + escapeHtml(data.message || fallbackText || '') + '</span>',
-    actions.length ? '<span class="export-dir-actions">' + actions.join('') + '</span>' : '',
-  ].join('');
-}
-
-function clearSettingsMessage() {
-  showSettingsMessage('', '');
-}
-
-function resetExportDirField() {
-  const input = document.getElementById('setExportDir');
-  if (!input) return;
-  input.value = '';
-  clearSettingsMessage();
-  renderExportDirStatus(null, '保存后将使用默认导出目录：' + (appSettings.export_dir_default || appSettings.export_dir_effective || '未记录'));
-  updateSettingsDirtyState();
-}
-
-function useDefaultExportDirFromError() {
-  resetExportDirField();
-  setOpsStatus('已切换为默认导出目录，保存后生效。', true);
-}
-
-function chooseExportDir() {
-  const input = document.getElementById('setExportDir');
-  const status = document.getElementById('exportDirStatus');
-  const button = document.getElementById('chooseExportDirButton');
-  const picker = window.pywebview && window.pywebview.api && window.pywebview.api.choose_export_dir;
-  if (!input) return;
-  clearSettingsMessage();
-  if (typeof picker !== 'function') {
-    const message = '当前浏览器模式不支持系统目录选择器，请手动输入导出目录。';
-    if (status) status.textContent = message;
-    setOpsStatus(message, false);
-    return;
-  }
-
-  if (status) status.textContent = '正在打开系统目录选择器...';
-  if (button) button.disabled = true;
-  Promise.resolve(picker.call(window.pywebview.api))
-    .then(data => {
-      const message = data && data.message ? data.message : '目录选择完成。';
-      if (!data || !data.ok) {
-        if (status) status.textContent = message;
-        if (!data || !data.cancelled) setOpsStatus(message, false);
-        return;
-      }
-      input.value = data.path || '';
-      const savedMessage = message + '，保存后生效。';
-      if (status) status.textContent = savedMessage;
-      setOpsStatus('已选择导出目录，保存后生效。', true);
-      updateSettingsDirtyState();
-    })
-    .catch(() => {
-      const message = '无法打开系统目录选择器，请手动输入导出目录。';
-      if (status) status.textContent = message;
-      setOpsStatus(message, false);
-    })
-    .finally(() => {
-      if (button) button.disabled = false;
-    });
-}
-
-function renderDeepseekModelOptions(selected) {
-  const select = document.getElementById('setDeepseekModel');
-  const model = selected || 'deepseek-v4-pro';
-  const options = Array.from(new Set([model, ...deepseekModelOptions].filter(Boolean)));
-  select.innerHTML = options.map(item => (
-    '<option value="' + escapeHtml(item) + '">' + escapeHtml(item) + '</option>'
-  )).join('');
-  select.value = model;
-}
-
-function updateRiskProviderFields() {
-  const provider = document.getElementById('setRiskAssistantProvider').value || 'deepseek';
-  document.querySelectorAll('.risk-provider-row').forEach(row => {
-    row.classList.toggle('hidden', row.getAttribute('data-provider') !== provider);
-  });
-}
-
-function refreshRiskModels() {
-  const status = document.getElementById('deepseekModelStatus');
-  if (status) {
-    status.textContent = '正在获取模型列表...';
-    delete status.dataset.state;
-  }
-  socket.emit('get_risk_model_options', { provider: 'deepseek' });
-}
-
-function testRiskModel() {
-  const status = document.getElementById('riskModelTestStatus');
-  if (status) {
-    status.textContent = '正在测试当前模型生成能力...';
-    status.className = 'model-test-status';
-  }
-  socket.emit('test_risk_model');
-}
-
-function settingsFieldElements() {
-  return SETTINGS_FIELD_IDS.map(id => document.getElementById(id)).filter(Boolean);
-}
-
-function captureSettingsSnapshot() {
-  const snapshot = {};
-  settingsFieldElements().forEach(element => {
-    snapshot[element.id] = element.type === 'checkbox' ? !!element.checked : element.value;
-  });
-  return JSON.stringify(snapshot);
-}
-
-function showSettingsMessage(message, state) {
-  const element = document.getElementById('settingsMessage');
-  if (!element) return;
-  element.textContent = message || '';
-  if (state) element.dataset.state = state;
-  else delete element.dataset.state;
-}
-
-function setSettingsSaving(saving) {
-  const button = document.getElementById('settingsSaveButton');
-  if (!button) return;
-  button.textContent = saving ? '正在保存' : '保存更改';
-  button.disabled = !!saving || !settingsDirty;
-}
-
-function setSettingsDirty(dirty) {
-  settingsDirty = !!dirty;
-  const state = document.getElementById('settingsDirtyState');
-  if (state) {
-    state.textContent = settingsDirty ? '有未保存的更改' : '所有更改已保存';
-    state.classList.toggle('changed', settingsDirty);
-  }
-  setSettingsSaving(pendingSettingsSave);
-}
-
-function updateSettingsDirtyState() {
-  if (!settingsInitialSnapshot) return;
-  setSettingsDirty(captureSettingsSnapshot() !== settingsInitialSnapshot);
-}
-
-function resetSettingsDirtySnapshot() {
-  settingsInitialSnapshot = captureSettingsSnapshot();
-  setSettingsDirty(false);
-  hideSettingsDiscardPrompt();
-}
-
-function readStoredSettingsTab() {
-  try {
-    const stored = window.localStorage.getItem(SETTINGS_TAB_STORAGE_KEY) || '';
-    return SETTINGS_TABS.includes(stored) ? stored : 'general';
-  } catch (_error) {
-    return 'general';
-  }
-}
-
-function storeSettingsTab(tab) {
-  try {
-    window.localStorage.setItem(SETTINGS_TAB_STORAGE_KEY, tab);
-  } catch (_error) {
-    // 本机浏览器禁用存储时，仅在当前会话中保留分页。
-  }
-}
-
-function settingsTabForElement(element) {
-  const panel = element && element.closest ? element.closest('[data-settings-panel]') : null;
-  return panel ? panel.dataset.settingsPanel : 'general';
-}
-
-function clearSettingsValidation() {
-  const modal = document.querySelector('.settings-primary-modal');
-  if (!modal) return;
-  modal.querySelectorAll('.invalid').forEach(element => element.classList.remove('invalid'));
-  modal.querySelectorAll('[aria-invalid="true"]').forEach(element => {
-    element.removeAttribute('aria-invalid');
-    const describedBy = element.getAttribute('aria-describedby') || '';
-    if (describedBy.endsWith('Error')) element.removeAttribute('aria-describedby');
-  });
-  modal.querySelectorAll('.setting-field-error').forEach(element => element.remove());
-  modal.querySelectorAll('.settings-tab.has-error').forEach(element => element.classList.remove('has-error'));
-}
-
-function clearSettingsFieldError(element) {
-  if (!element) return;
-  element.classList.remove('invalid');
-  const inlineCheck = element.closest('.inline-check');
-  if (inlineCheck) inlineCheck.classList.remove('invalid');
-  element.removeAttribute('aria-invalid');
-  const error = document.getElementById(element.id + 'Error');
-  if (error) error.remove();
-  const tab = settingsTabForElement(element);
-  const panel = document.querySelector('[data-settings-panel="' + tab + '"]');
-  if (panel && !panel.querySelector('.setting-field-error')) {
-    const tabButton = document.querySelector('[data-settings-tab="' + tab + '"]');
-    if (tabButton) tabButton.classList.remove('has-error');
-  }
-}
-
-function renderSettingsValidation(errors) {
-  clearSettingsValidation();
-  errors.forEach(error => {
-    const element = document.getElementById(error.id);
-    if (!element) return;
-    const highlight = element.closest('.inline-check') || element;
-    highlight.classList.add('invalid');
-    element.setAttribute('aria-invalid', 'true');
-    const row = element.closest('.setting-row');
-    if (row && !document.getElementById(error.id + 'Error')) {
-      const note = document.createElement('div');
-      note.className = 'setting-field-error';
-      note.id = error.id + 'Error';
-      note.textContent = error.message;
-      row.appendChild(note);
-      element.setAttribute('aria-describedby', note.id);
-    }
-    const tabButton = document.querySelector('[data-settings-tab="' + error.tab + '"]');
-    if (tabButton) tabButton.classList.add('has-error');
-  });
-}
-
-function validateSettings() {
-  const errors = [];
-  const add = (id, message) => {
-    const element = document.getElementById(id);
-    if (!element) return;
-    errors.push({ id, message, tab: settingsTabForElement(element) });
-  };
-  const validateNumber = (id, label, min, max) => {
-    const element = document.getElementById(id);
-    if (!element || element.value.trim() === '') return;
-    const value = Number(element.value);
-    if (!Number.isFinite(value) || value < min || value > max) add(id, label + '应在 ' + min + ' 到 ' + max + ' 之间。');
-  };
-  const validateTypedValue = (id, message) => {
-    const element = document.getElementById(id);
-    if (element && element.value.trim() && !element.validity.valid) add(id, message);
-  };
-
-  validateNumber('setFloatingOpacity', '悬浮条透明度', 50, 100);
-  validateNumber('setAlertCooldownMinutes', '提醒冷却时间', 0, 240);
-  validateNumber('setSmtpPort', 'SMTP 端口', 1, 65535);
-  validateNumber('setRiskMaxTokens', '单次输出上限', 300, 4000);
-  validateNumber('setRiskCooldownSeconds', '分析冷却时间', 0, 300);
-  validateNumber('setRiskCacheMinutes', '重复分析缓存', 0, 60);
-  validateTypedValue('setSmtpSender', '发件邮箱格式不正确。');
-  validateTypedValue('setSmtpRecipient', '收件邮箱格式不正确。');
-  validateTypedValue('setWebhookUrl', 'Webhook 地址格式不正确。');
-  validateTypedValue('setDeepseekBaseUrl', 'DeepSeek API 地址格式不正确。');
-  validateTypedValue('setOpenaiCompatibleBaseUrl', '兼容接口地址格式不正确。');
-
-  const quietStart = document.getElementById('setAlertQuietStart').value;
-  const quietEnd = document.getElementById('setAlertQuietEnd').value;
-  if (!!quietStart !== !!quietEnd) add(quietStart ? 'setAlertQuietEnd' : 'setAlertQuietStart', '静默时段需要同时填写开始和结束时间。');
-
-  if (document.getElementById('setWebhookEnabled').checked) {
-    const webhookUrl = document.getElementById('setWebhookUrl').value.trim();
-    if (!webhookUrl) add('setWebhookUrl', '启用 Webhook 前需要填写接收地址。');
-    else if (!webhookUrl.toLowerCase().startsWith('https://')) add('setWebhookUrl', 'Webhook 地址必须使用 HTTPS。');
-  }
-
-  if (document.getElementById('setDailyDigestEnabled').checked) {
-    if (!document.getElementById('setDailyDigestTime').value) add('setDailyDigestTime', '启用每日摘要前需要设置发送时间。');
-    if (!document.getElementById('setDailyDigestEmail').checked && !document.getElementById('setDailyDigestWebhook').checked) {
-      add('setDailyDigestEmail', '启用每日摘要前至少选择一个发送渠道。');
-    }
-  }
-
-  if (document.getElementById('setRiskAssistantEnabled').checked) {
-    const provider = document.getElementById('setRiskAssistantProvider').value;
-    if (provider === 'deepseek') {
-      if (!document.getElementById('setDeepseekBaseUrl').value.trim()) add('setDeepseekBaseUrl', '需要填写 DeepSeek API 地址。');
-      if (!document.getElementById('setDeepseekModel').value) add('setDeepseekModel', '需要选择 DeepSeek 模型。');
-    } else {
-      if (!document.getElementById('setOpenaiCompatibleBaseUrl').value.trim()) add('setOpenaiCompatibleBaseUrl', '需要填写兼容接口地址。');
-      if (!document.getElementById('setOpenaiCompatibleModel').value.trim()) add('setOpenaiCompatibleModel', '需要填写兼容模型名称。');
-    }
-  }
-
-  renderSettingsValidation(errors);
-  if (!errors.length) return true;
-  const first = errors[0];
-  switchSettingsTab(first.tab);
-  showSettingsMessage(
-    SETTINGS_TAB_LABELS[first.tab] + '有 ' + errors.filter(error => error.tab === first.tab).length + ' 项需要处理：' + first.message,
-    'error'
-  );
-  window.requestAnimationFrame(() => {
-    const element = document.getElementById(first.id);
-    if (element) element.focus();
-  });
-  return false;
-}
-
-function handleSettingsFieldChange(event) {
-  const target = event.target;
-  if (!target || !SETTINGS_FIELD_IDS.includes(target.id)) return;
-  clearSettingsFieldError(target);
-  if (target.id === 'setDailyDigestEmail' || target.id === 'setDailyDigestWebhook') {
-    clearSettingsFieldError(document.getElementById('setDailyDigestEmail'));
-  }
-  if (target.id === 'setAlertQuietStart' || target.id === 'setAlertQuietEnd') {
-    clearSettingsFieldError(document.getElementById('setAlertQuietStart'));
-    clearSettingsFieldError(document.getElementById('setAlertQuietEnd'));
-  }
-  const message = document.getElementById('settingsMessage');
-  if (message && message.dataset.state === 'error') showSettingsMessage('', '');
-  updateSettingsDirtyState();
-}
-
-function handleSettingsTabKeydown(event) {
-  const current = event.target.closest && event.target.closest('.settings-tab');
-  if (!current) return;
-  const tabs = Array.from(document.querySelectorAll('.settings-tab'));
-  const index = tabs.indexOf(current);
-  if (index < 0) return;
-  let nextIndex = index;
-  if (event.key === 'ArrowDown' || event.key === 'ArrowRight') nextIndex = (index + 1) % tabs.length;
-  else if (event.key === 'ArrowUp' || event.key === 'ArrowLeft') nextIndex = (index - 1 + tabs.length) % tabs.length;
-  else if (event.key === 'Home') nextIndex = 0;
-  else if (event.key === 'End') nextIndex = tabs.length - 1;
-  else return;
-  event.preventDefault();
-  const next = tabs[nextIndex];
-  switchSettingsTab(next.dataset.settingsTab);
-  next.focus();
-}
-
-function switchSettingsTab(tab) {
-  const nextTab = SETTINGS_TABS.includes(tab) ? tab : 'general';
-  activeSettingsTab = nextTab;
-  SETTINGS_TABS.forEach(name => {
-    const active = nextTab === name;
-    const suffix = name.charAt(0).toUpperCase() + name.slice(1);
-    const tabButton = document.getElementById('settingsTab' + suffix);
-    const panel = document.getElementById('settingsPanel' + suffix);
-    tabButton.classList.toggle('active', active);
-    tabButton.setAttribute('aria-selected', String(active));
-    tabButton.tabIndex = active ? 0 : -1;
-    panel.classList.toggle('active', active);
-    panel.hidden = !active;
-  });
-  storeSettingsTab(nextTab);
-  const activeButton = document.querySelector('[data-settings-tab="' + nextTab + '"]');
-  if (activeButton) {
-    window.requestAnimationFrame(() => activeButton.scrollIntoView({ block: 'nearest', inline: 'nearest' }));
-  }
-  const body = document.querySelector('#settingsBackdrop .settings-body');
-  if (body) body.scrollTop = 0;
-  if (nextTab === 'risk') refreshRiskModels();
-  if (nextTab === 'digest') socket.emit('get_daily_digest_status');
-}
-
-function openSettings() {
-  settingsLastFocused = document.activeElement;
-  applySettings(appSettings);
-  clearSettingsValidation();
-  showSettingsMessage('', '');
-  document.getElementById('settingsBackdrop').classList.add('show');
-  switchSettingsTab(readStoredSettingsTab());
-  resetSettingsDirtySnapshot();
-  window.requestAnimationFrame(() => {
-    const activeTab = document.querySelector('.settings-tab.active');
-    if (activeTab) activeTab.focus();
-  });
-}
-
-function onboardingPriceText() {
-  if (Number.isFinite(Number(latestData.rmb))) return '¥' + Number(latestData.rmb).toFixed(2) + '/克';
-  if (Number.isFinite(Number(latestData.usd))) return '$' + Number(latestData.usd).toFixed(2) + '/盎司';
-  return '等待行情';
-}
-
-function updateOnboardingStatus() {
-  const market = document.getElementById('onboardingMarketStatus');
-  const price = document.getElementById('onboardingPriceStatus');
-  const source = document.getElementById('onboardingSourceStatus');
-  const statusText = document.getElementById('statusText');
-  if (market) market.textContent = statusText && statusText.textContent ? statusText.textContent : '本地服务已连接';
-  if (price) price.textContent = onboardingPriceText();
-  if (source) source.textContent = latestData.gold_source || '等待行情';
-}
-
-function populateOnboardingFields() {
-  document.getElementById('onboardingDisplayMode').value = appSettings.floating_price_display_mode || 'rmb_usd';
-  document.getElementById('onboardingFloatingEnabled').checked = appSettings.floating_price_enabled !== false;
-  document.getElementById('onboardingStartupEnabled').checked = !!appSettings.startup_enabled;
-  document.getElementById('onboardingStartupTray').checked = appSettings.startup_to_tray !== false;
-  document.getElementById('onboardingCloseBehavior').value = appSettings.close_behavior || 'ask';
-  document.getElementById('onboardingAlertSound').checked = appSettings.alert_sound_enabled !== false;
-  document.getElementById('onboardingAlertDialog').checked = appSettings.alert_dialog_enabled !== false;
-  const cooldown = document.getElementById('onboardingCooldown');
-  const cooldownValue = String(appSettings.alert_cooldown_minutes ?? 30);
-  if (![...cooldown.options].some(option => option.value === cooldownValue)) {
-    const option = document.createElement('option');
-    option.value = cooldownValue;
-    option.textContent = cooldownValue + ' 分钟';
-    cooldown.appendChild(option);
-  }
-  cooldown.value = cooldownValue;
-}
-
-function onboardingPreferences() {
-  return {
-    floating_price_display_mode: document.getElementById('onboardingDisplayMode').value,
-    floating_price_enabled: document.getElementById('onboardingFloatingEnabled').checked,
-    startup_enabled: document.getElementById('onboardingStartupEnabled').checked,
-    startup_to_tray: document.getElementById('onboardingStartupTray').checked,
-    close_behavior: document.getElementById('onboardingCloseBehavior').value,
-    alert_sound_enabled: document.getElementById('onboardingAlertSound').checked,
-    alert_dialog_enabled: document.getElementById('onboardingAlertDialog').checked,
-    alert_cooldown_minutes: document.getElementById('onboardingCooldown').value,
-  };
-}
-
-function renderOnboardingSummary() {
-  const summary = document.getElementById('onboardingSummary');
-  if (!summary) return;
-  const preferences = onboardingPreferences();
-  const displayLabels = { rmb_usd: '人民币与美元', rmb_only: '仅人民币', usd_only: '仅美元' };
-  const closeLabels = { ask: '关闭时询问', minimize_to_tray: '继续在后台运行', exit: '退出程序' };
-  const parts = [
-    '悬浮窗：' + (preferences.floating_price_enabled ? displayLabels[preferences.floating_price_display_mode] : '不启用'),
-    '开机自启动：' + (preferences.startup_enabled ? '启用' : '不启用'),
-    '关闭行为：' + (closeLabels[preferences.close_behavior] || '关闭时询问'),
-    '提示音：' + (preferences.alert_sound_enabled ? '启用' : '关闭'),
-    '警报窗口：' + (preferences.alert_dialog_enabled ? '启用' : '关闭'),
-    '相同规则冷却：' + preferences.alert_cooldown_minutes + ' 分钟',
-  ];
-  summary.innerHTML = parts.map(part => '<div>' + escapeHtml(part) + '</div>').join('');
-}
-
-function showOnboardingStep(step) {
-  onboardingStep = Math.max(1, Math.min(4, Number(step) || 1));
-  document.querySelectorAll('[data-onboarding-step]').forEach(section => {
-    section.classList.toggle('active', Number(section.dataset.onboardingStep) === onboardingStep);
-  });
-  document.querySelectorAll('[data-onboarding-progress]').forEach(item => {
-    item.classList.toggle('active', Number(item.dataset.onboardingProgress) <= onboardingStep);
-  });
-  document.getElementById('onboardingBackButton').hidden = onboardingStep === 1;
-  document.getElementById('onboardingNextButton').hidden = onboardingStep === 4;
-  document.getElementById('onboardingFinishButton').hidden = onboardingStep !== 4;
-  if (onboardingStep === 1) updateOnboardingStatus();
-  if (onboardingStep === 4) renderOnboardingSummary();
-}
-
-function openOnboarding(manual) {
-  onboardingManual = !!manual;
-  populateOnboardingFields();
-  const message = document.getElementById('onboardingMessage');
-  if (message) message.textContent = '';
-  document.getElementById('onboardingSkipButton').textContent = onboardingManual ? '关闭向导' : '暂不设置';
-  document.getElementById('onboardingBackdrop').classList.add('show');
-  showOnboardingStep(1);
-  if (!onboardingManual && !appSettings.onboarding_started) socket.emit('start_onboarding');
-}
-
-function maybeOpenOnboarding() {
-  if (onboardingAutoChecked) return;
-  onboardingAutoChecked = true;
-  if (appSettings.onboarding_completed) return;
-  setTimeout(() => openOnboarding(false), 120);
-}
-
-function reopenOnboarding() {
-  if (settingsDirty || pendingSettingsSave) {
-    showSettingsMessage('请先保存或放弃当前更改，再重新打开首次使用向导。', 'error');
-    return;
-  }
-  closeSettings(true);
-  openOnboarding(true);
-}
-
-function changeOnboardingStep(delta) {
-  showOnboardingStep(onboardingStep + Number(delta || 0));
-}
-
-function finishOnboarding() {
-  const finishButton = document.getElementById('onboardingFinishButton');
-  if (finishButton) finishButton.disabled = true;
-  socket.emit('complete_onboarding', onboardingPreferences());
-}
-
-function skipOnboarding() {
-  if (onboardingManual) {
-    document.getElementById('onboardingBackdrop').classList.remove('show');
-    return;
-  }
-  const skipButton = document.getElementById('onboardingSkipButton');
-  if (skipButton) skipButton.disabled = true;
-  socket.emit('complete_onboarding', {});
-}
-
-function hideSettingsDiscardPrompt() {
-  const prompt = document.getElementById('settingsUnsavedConfirm');
-  if (prompt) prompt.hidden = true;
-}
-
-function discardSettingsChanges() {
-  closeSettings(true);
-}
-
-function settingsFocusableElements() {
-  const modal = document.querySelector('.settings-primary-modal');
-  if (!modal) return [];
-  return Array.from(modal.querySelectorAll('button:not([disabled]), input:not([disabled]), select:not([disabled]), textarea:not([disabled]), [tabindex]:not([tabindex="-1"])'))
-    .filter(element => !element.hidden && element.offsetParent !== null);
-}
-
-function handleSettingsDialogKeydown(event) {
-  const backdrop = document.getElementById('settingsBackdrop');
-  if (!backdrop || !backdrop.classList.contains('show')) return;
-  if (event.key === 'Escape') {
-    event.preventDefault();
-    closeSettings();
-    return;
-  }
-  if (event.key !== 'Tab') return;
-  const focusable = settingsFocusableElements();
-  if (!focusable.length) return;
-  const first = focusable[0];
-  const last = focusable[focusable.length - 1];
-  if (event.shiftKey && document.activeElement === first) {
-    event.preventDefault();
-    last.focus();
-  } else if (!event.shiftKey && document.activeElement === last) {
-    event.preventDefault();
-    first.focus();
-  }
-}
-
-function closeSettings(force) {
-  const backdrop = document.getElementById('settingsBackdrop');
-  if (!backdrop || !backdrop.classList.contains('show')) return;
-  if (!force && pendingSettingsSave) {
-    showSettingsMessage('设置正在保存，请等待后台确认。', '');
-    return;
-  }
-  if (!force && settingsDirty) {
-    const prompt = document.getElementById('settingsUnsavedConfirm');
-    if (prompt) prompt.hidden = false;
-    const discardButton = document.getElementById('settingsDiscardButton');
-    if (discardButton) discardButton.focus();
-    return;
-  }
-  backdrop.classList.remove('show');
-  hideSettingsDiscardPrompt();
-  clearSettingsValidation();
-  settingsInitialSnapshot = '';
-  settingsDirty = false;
-  if (settingsLastFocused && typeof settingsLastFocused.focus === 'function') settingsLastFocused.focus();
-  settingsLastFocused = null;
-}
-
-function onSettingsBackdrop(event) {
-  if (event.target.id === 'settingsBackdrop') closeSettings();
-}
-
-function setupSettingsInteractions() {
-  const backdrop = document.getElementById('settingsBackdrop');
-  const tabs = document.querySelector('.settings-primary-modal .settings-tabs');
-  if (backdrop) {
-    backdrop.addEventListener('input', handleSettingsFieldChange);
-    backdrop.addEventListener('change', handleSettingsFieldChange);
-  }
-  if (tabs) tabs.addEventListener('keydown', handleSettingsTabKeydown);
-  document.addEventListener('keydown', handleSettingsDialogKeydown);
-}
-
-setupSettingsInteractions();
-
-function saveSettings() {
-  if (pendingSettingsSave) return;
-  if (!settingsDirty) {
-    showSettingsMessage('当前没有需要保存的更改。', '');
-    return;
-  }
-  if (!validateSettings()) return;
-  const closeBehavior = document.getElementById('setCloseBehavior').value;
-  const next = {
-    startup_enabled: document.getElementById('setStartup').checked,
-    startup_to_tray: document.getElementById('setStartupTray').checked,
-    floating_price_enabled: document.getElementById('setFloatingPrice').checked,
-    floating_price_display_mode: document.getElementById('setFloatingDisplayMode').value,
-    floating_price_preset: document.getElementById('setFloatingPreset').value,
-    floating_price_opacity: document.getElementById('setFloatingOpacity').value.trim(),
-    floating_price_snap_edge: document.getElementById('setFloatingSnapEdge').checked,
-    floating_price_always_on_top: document.getElementById('setFloatingAlwaysOnTop').checked,
-    close_behavior: closeBehavior,
-    close_remembered: closeBehavior !== 'ask',
-    alert_sound_enabled: document.getElementById('setAlertSound').checked,
-    alert_dialog_enabled: document.getElementById('setAlertDialog').checked,
-    // 邮件通知
-    smtp_server: document.getElementById('setSmtpServer').value.trim(),
-    smtp_port: document.getElementById('setSmtpPort').value.trim(),
-    smtp_encryption: document.getElementById('setSmtpEncryption').value,
-    smtp_sender: document.getElementById('setSmtpSender').value.trim(),
-    smtp_password: document.getElementById('setSmtpPassword').value,
-    smtp_password_clear: document.getElementById('clearSmtpPassword').checked,
-    smtp_recipient: document.getElementById('setSmtpRecipient').value.trim(),
-    email_warning_enabled: appSettings.email_warning_enabled !== false,
-    email_critical_enabled: appSettings.email_critical_enabled !== false,
-    email_volatility_enabled: appSettings.email_volatility_enabled !== false,
-    alert_cooldown_minutes: document.getElementById('setAlertCooldownMinutes').value.trim(),
-    alert_quiet_start: document.getElementById('setAlertQuietStart').value,
-    alert_quiet_end: document.getElementById('setAlertQuietEnd').value,
-    email_subject_template: document.getElementById('setEmailSubjectTemplate').value,
-    email_body_template: document.getElementById('setEmailBodyTemplate').value,
-    webhook_enabled: document.getElementById('setWebhookEnabled').checked,
-    webhook_url: document.getElementById('setWebhookUrl').value.trim(),
-    webhook_warning_enabled: document.getElementById('setWebhookWarning').checked,
-    webhook_critical_enabled: document.getElementById('setWebhookCritical').checked,
-    webhook_volatility_enabled: document.getElementById('setWebhookVolatility').checked,
-    daily_digest_enabled: document.getElementById('setDailyDigestEnabled').checked,
-    daily_digest_time: document.getElementById('setDailyDigestTime').value,
-    daily_digest_email_enabled: document.getElementById('setDailyDigestEmail').checked,
-    daily_digest_webhook_enabled: document.getElementById('setDailyDigestWebhook').checked,
-    risk_assistant_enabled: document.getElementById('setRiskAssistantEnabled').checked,
-    risk_assistant_provider: document.getElementById('setRiskAssistantProvider').value,
-    risk_assistant_depth: document.getElementById('setRiskAssistantDepth').value,
-    deepseek_base_url: document.getElementById('setDeepseekBaseUrl').value.trim(),
-    deepseek_model: document.getElementById('setDeepseekModel').value,
-    deepseek_api_key: document.getElementById('setDeepseekApiKey').value.trim(),
-    deepseek_api_key_clear: document.getElementById('clearDeepseekApiKey').checked,
-    openai_compatible_base_url: document.getElementById('setOpenaiCompatibleBaseUrl').value.trim(),
-    openai_compatible_model: document.getElementById('setOpenaiCompatibleModel').value.trim(),
-    openai_compatible_api_key: document.getElementById('setOpenaiCompatibleApiKey').value.trim(),
-    openai_compatible_api_key_clear: document.getElementById('clearOpenaiCompatibleApiKey').checked,
-    risk_assistant_max_tokens: document.getElementById('setRiskMaxTokens').value.trim(),
-    risk_assistant_cooldown_seconds: document.getElementById('setRiskCooldownSeconds').value.trim(),
-    risk_assistant_cache_minutes: document.getElementById('setRiskCacheMinutes').value.trim(),
-    export_dir: document.getElementById('setExportDir').value.trim(),
-  };
-  pendingSettingsSave = true;
-  settingsSaveFailed = false;
-  setSettingsSaving(true);
-  showSettingsMessage('正在保存并应用设置...', '');
-  socket.emit('update_settings', next);
-  if (settingsSaveTimer) clearTimeout(settingsSaveTimer);
-  settingsSaveTimer = setTimeout(() => {
-    if (!pendingSettingsSave) return;
-    pendingSettingsSave = false;
-    settingsSaveFailed = true;
-    setSettingsDirty(true);
-    showSettingsMessage('保存失败：后台服务未响应，请退出托盘中的旧程序后重新打开最新版。', 'error');
-  }, 5000);
-}
-
-function testEmail() {
-  const statusEl = document.getElementById('testEmailStatus');
-  const btn = document.getElementById('btnTestEmail');
-  statusEl.textContent = '正在发送...';
-  statusEl.className = 'test-email-status';
-  btn.disabled = true;
-  socket.emit('test_email');
-}
-
-socket.on('test_email_result', data => {
-  const statusEl = document.getElementById('testEmailStatus');
-  const btn = document.getElementById('btnTestEmail');
-  btn.disabled = false;
-  if (data.ok) {
-    statusEl.textContent = data.message;
-    statusEl.className = 'test-email-status ok';
-  } else {
-    statusEl.textContent = data.message;
-    statusEl.className = 'test-email-status fail';
-  }
-});
-
-function testWebhook() {
-  const statusEl = document.getElementById('testWebhookStatus');
-  const btn = document.getElementById('btnTestWebhook');
-  if (statusEl) {
-    statusEl.textContent = '正在发送...';
-    statusEl.className = 'test-email-status';
-  }
-  if (btn) btn.disabled = true;
-  socket.emit('test_webhook');
-}
-
-socket.on('test_webhook_result', data => {
-  const statusEl = document.getElementById('testWebhookStatus');
-  const btn = document.getElementById('btnTestWebhook');
-  if (btn) btn.disabled = false;
-  if (!statusEl) return;
-  statusEl.textContent = data && data.message ? data.message : 'Webhook 测试完成。';
-  statusEl.className = 'test-email-status ' + (data && data.ok ? 'ok' : 'fail');
-});
-
-function previewDailyDigest() {
-  const button = document.getElementById('btnPreviewDailyDigest');
-  if (button) button.disabled = true;
-  setDailyDigestStatus('正在生成摘要预览...', null);
-  socket.emit('preview_daily_digest');
-}
-
-function testDailyDigest() {
-  const button = document.getElementById('btnTestDailyDigest');
-  if (button) button.disabled = true;
-  setDailyDigestStatus('正在测试发送每日摘要...', null);
-  socket.emit('test_daily_digest');
-}
-
-function testAlert() {
-  const statusEl = document.getElementById('testAlertStatus');
-  const btn = document.getElementById('btnTestAlert');
-  if (statusEl) {
-    statusEl.textContent = '正在触发...';
-    statusEl.className = 'test-email-status';
-  }
-  if (btn) {
-    btn.disabled = true;
-    setTimeout(() => { btn.disabled = false; }, 1200);
-  }
-  socket.emit('test_alert', { type: 'warning' });
 }
 
 // ========== 通用界面工具 ==========
@@ -2783,130 +868,6 @@ function setupEllipsisTooltips() {
 }
 
 setupEllipsisTooltips();
-
-// ========== 更新 ==========
-function openUpdate() {
-  document.getElementById('updateBackdrop').classList.add('show');
-  if (!pendingUpdateInfo) {
-    checkUpdate();
-  }
-}
-
-function closeUpdate() {
-  document.getElementById('updateBackdrop').classList.remove('show');
-}
-
-function onUpdateBackdrop(event) {
-  if (event.target.id === 'updateBackdrop') closeUpdate();
-}
-
-function renderOpsUpdateStatus(data) {
-  opsUpdateStatus = data || opsUpdateStatus || null;
-  const statusEl = document.getElementById('opsUpdateStatus');
-  const metaEl = document.getElementById('opsUpdateMeta');
-  if (!statusEl || !metaEl) return;
-  const state = opsUpdateStatus && opsUpdateStatus.state ? opsUpdateStatus.state : '';
-  const message = opsUpdateStatus && opsUpdateStatus.message ? opsUpdateStatus.message : '尚未检查更新。';
-  const current = opsUpdateStatus && opsUpdateStatus.current_version ? '当前版本 ' + opsUpdateStatus.current_version : '';
-  const latest = opsUpdateStatus && opsUpdateStatus.latest_version ? '最新版本 ' + opsUpdateStatus.latest_version : '';
-  const checked = opsUpdateStatus && opsUpdateStatus.checked_at ? '检查时间 ' + String(opsUpdateStatus.checked_at).replace('T', ' ') : '';
-  statusEl.textContent = message;
-  statusEl.dataset.state = state || 'unknown';
-  const meta = [current, latest, checked].filter(Boolean).join(' · ');
-  if (meta) metaEl.textContent = meta;
-}
-
-function checkUpdateFromOps() {
-  renderOpsUpdateStatus({ state: 'checking', message: '正在检查更新...' });
-  requestUpdateCheck(true);
-  setOpsStatus('正在检查更新...', true);
-}
-
-function openUpdateFromOps() {
-  openUpdate();
-}
-
-function checkUpdate() {
-  requestUpdateCheck(false);
-}
-
-function requestUpdateCheck(silent) {
-  pendingUpdateInfo = null;
-  document.getElementById('updateButton').classList.remove('update-ready');
-  document.getElementById('installUpdateButton').disabled = true;
-  renderOpsUpdateStatus({ state: 'checking', message: '正在检查更新...' });
-  if (!silent) {
-    document.getElementById('updateStatus').textContent = '正在检查更新...';
-    document.getElementById('updateMeta').textContent = '';
-    document.getElementById('updateNotes').style.display = 'none';
-  }
-  lastAutoUpdateCheckAt = Date.now();
-  socket.emit('check_update');
-}
-
-function scheduleAutoUpdateCheck() {
-  if (autoUpdateTimer) {
-    clearTimeout(autoUpdateTimer);
-    autoUpdateTimer = null;
-  }
-  const elapsed = Date.now() - lastAutoUpdateCheckAt;
-  const delay = lastAutoUpdateCheckAt ? Math.max(autoUpdateIntervalMs() - elapsed, 60 * 1000) : 2000;
-  autoUpdateTimer = setTimeout(() => {
-    requestUpdateCheck(true);
-    scheduleAutoUpdateCheck();
-  }, delay);
-}
-
-function applyUpdateStatus(data) {
-  renderOpsUpdateStatus(data);
-  const statusEl = document.getElementById('updateStatus');
-  const metaEl = document.getElementById('updateMeta');
-  const notesEl = document.getElementById('updateNotes');
-  const installBtn = document.getElementById('installUpdateButton');
-  const updateBtn = document.getElementById('updateButton');
-  const progressEl = document.getElementById('updateProgress');
-  const progressBar = progressEl ? progressEl.querySelector('span') : null;
-
-  const current = data.current_version ? '当前版本 ' + data.current_version : '';
-  const latest = data.latest_version ? '最新版本 ' + data.latest_version : '';
-  const checked = data.checked_at ? '检查时间 ' + String(data.checked_at).replace('T', ' ') : '';
-  metaEl.textContent = [current, latest, checked].filter(Boolean).join(' · ');
-  statusEl.textContent = data.message || '更新状态未知。';
-  notesEl.textContent = data.notes || '';
-  notesEl.style.display = data.notes ? 'block' : 'none';
-  if (progressEl && progressBar) {
-    const percent = data.progress_percent == null ? null : Math.max(0, Math.min(100, Number(data.progress_percent)));
-    progressEl.classList.toggle('show', data.state === 'downloading' || data.state === 'installing');
-    progressBar.style.width = Number.isFinite(percent) ? percent + '%' : '35%';
-  }
-
-  if (data.state === 'available') {
-    pendingUpdateInfo = {
-      version: data.latest_version,
-      notes: data.notes || '',
-    };
-    installBtn.disabled = false;
-    updateBtn.classList.add('update-ready');
-    return;
-  }
-
-  if (data.state === 'downloading' || data.state === 'installing') {
-    installBtn.disabled = true;
-    updateBtn.classList.add('update-ready');
-    return;
-  }
-
-  pendingUpdateInfo = null;
-  installBtn.disabled = true;
-  updateBtn.classList.remove('update-ready');
-}
-
-function installUpdate() {
-  if (!pendingUpdateInfo) return;
-  document.getElementById('installUpdateButton').disabled = true;
-  document.getElementById('updateStatus').textContent = '正在准备更新...';
-  socket.emit('install_update');
-}
 
 // ========== 行情状态 ==========
 function applyFetchStatus(data) {
