+++ templates/index.html
@@ -673,6 +673,8 @@
 <script src="https://cdn.socket.io/4.7.2/socket.io.min.js"></script>
 <script src="/static/app-shell.js?v={{ app_version }}"></script>
 
+<script src="/static/settings-center.js?v={{ app_version }}"></script>
+<script src="/static/operations-center.js?v={{ app_version }}"></script>
 <script src="/static/history-review-center.js?v={{ app_version }}"></script>
 <script src="/static/risk-analysis-center.js?v={{ app_version }}"></script>
 <script src="/static/alert-rule-center.js?v={{ app_version }}"></script>
