+++ tests/test_run_checks_script.py
@@ -15,8 +15,10 @@ EXPECTED_COMPILE_TARGETS = (
     "goldmonitor/scheduler.py",
     "goldmonitor/socket_alert_rules.py",
     "goldmonitor/socket_history_review.py",
+    "goldmonitor/socket_operations.py",
     "goldmonitor/socket_portfolio.py",
     "goldmonitor/socket_risk_analysis.py",
+    "goldmonitor/socket_settings.py",
     "tests/risk_contract_check.py",
     "tests/update_logic_check.py",
     "tests/startup_contract_check.py",
